+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix an issue with checkbox list items separated with blank lines (#602).
 * Require package `web: '>=0.4.2 <2.0.0'`.
+* #16: `Document` supports the preserveSpace option
 * `parseInlineSyntax': add record type annotation to the return value
 
 ## 7.2.2